Move lru_cache to utils for reuse
