+++ src/pip/_internal/utils/compat.py
@@ -7,6 +7,7 @@ distributions."""
 from __future__ import absolute_import, division
 
 import codecs
+import functools
 import locale
 import logging
 import os
@@ -18,7 +19,15 @@ from pip._vendor.six import PY2, text_type
 from pip._internal.utils.typing import MYPY_CHECK_RUNNING
 
 if MYPY_CHECK_RUNNING:
-    from typing import Optional, Text, Tuple, Union
+    from typing import Callable, Optional, Protocol, Text, Tuple, TypeVar, Union
+
+    # Used in the @lru_cache polyfill.
+    F = TypeVar('F')
+
+    class LruCache(Protocol):
+        def __call__(self, maxsize=None):
+            # type: (Optional[int]) -> Callable[[F], F]
+            raise NotImplementedError
 
 try:
     import ipaddress
@@ -269,3 +278,16 @@ else:
         if not cr:
             cr = (os.environ.get('LINES', 25), os.environ.get('COLUMNS', 80))
         return int(cr[1]), int(cr[0])
+
+
+# Fallback to noop_lru_cache in Python 2
+# TODO: this can be removed when python 2 support is dropped!
+def noop_lru_cache(maxsize=None):
+    # type: (Optional[int]) -> Callable[[F], F]
+    def _wrapper(f):
+        # type: (F) -> F
+        return f
+    return _wrapper
+
+
+lru_cache = getattr(functools, "lru_cache", noop_lru_cache)  # type: LruCache
